refactor: modularize COREMOD_arith execution engines and domain functions
