+++ src/coremods/COREMOD_arith/imfunctions.c
@@ -922,131 +922,7 @@ errno_t arith_image_function_2_1_inplace(
 /* ------------------------------------------------------------------------- */
 /* complex image, complex image  -> complex image                            */
 /* ------------------------------------------------------------------------- */
-// complex float (CF), complex float (CF) -> complex float (CF)
-errno_t arith_image_function_CF_CF__CF(
-    const char *ID_name1,
-    const char *ID_name2,
-    const char *ID_out,
-    complex_float(*pt2function)(
-        complex_float, complex_float))
-{
-    IMGID img1 =
-        imgid_make_from_name(ID_name1);
-    resolveIMGID(&img1,
-                 ERRMODE_ABORT,
-                 dcimg, dcnimg);
-
-    IMGID img2 =
-        imgid_make_from_name(ID_name2);
-    resolveIMGID(&img2,
-                 ERRMODE_ABORT,
-                 dcimg, dcnimg);
-
-    IMGID imgout =
-        imgid_make_from_name(ID_out);
-    imgout.mdt->naxis = img1.md->naxis;
-    for(long i = 0;
-        i < img1.md->naxis; i++)
-    {
-        imgout.mdt->size[i] =
-            img1.md->size[i];
-    }
-    imgout.mdt->datatype =
-        img1.md->datatype;
-    imgout.mdt->shared = dcshareddft;
-    imgout.mdt->NBkw = NB_KEYWNODE_MAX;
-    imgout.im = (IMAGE *) calloc(
-        1, sizeof(IMAGE));
-    imgid_mkimage(&imgout);
-
-    long nelement =
-        img1.md->nelement;
-
-#ifdef _OPENMP
-    #pragma omp parallel if (nelement > OMP_NELEMENT_LIMIT)
-    {
-        #pragma omp for simd
-#endif
-        for(long ii = 0;
-            ii < nelement; ii++)
-        {
-            imgout.im->array.CF[ii] =
-                pt2function(
-                    img1.im->array.CF[ii],
-                    img2.im->array.CF[ii]);
-        }
-#ifdef _OPENMP
-    }
-#endif
-    RegisterIMGID(&imgout, dcimg, dcnimg);
-    imgid_free(&img1);
-    imgid_free(&img2);
-    imgid_free(&imgout);
-    return RETURN_SUCCESS;
-}
-
-// complex double (CD), complex double (CD) -> complex double (CD)
-errno_t arith_image_function_CD_CD__CD(
-    const char *ID_name1,
-    const char *ID_name2,
-    const char *ID_out,
-    complex_double(*pt2function)(
-        complex_double, complex_double))
-{
-    IMGID img1 =
-        imgid_make_from_name(ID_name1);
-    resolveIMGID(&img1,
-                 ERRMODE_ABORT,
-                 dcimg, dcnimg);
-
-    IMGID img2 =
-        imgid_make_from_name(ID_name2);
-    resolveIMGID(&img2,
-                 ERRMODE_ABORT,
-                 dcimg, dcnimg);
-
-    IMGID imgout =
-        imgid_make_from_name(ID_out);
-    imgout.mdt->naxis = img1.md->naxis;
-    for(long i = 0;
-        i < img1.md->naxis; i++)
-    {
-        imgout.mdt->size[i] =
-            img1.md->size[i];
-    }
-    imgout.mdt->datatype =
-        img1.md->datatype;
-    imgout.mdt->shared = dcshareddft;
-    imgout.mdt->NBkw = NB_KEYWNODE_MAX;
-    imgout.im = (IMAGE *) calloc(
-        1, sizeof(IMAGE));
-    imgid_mkimage(&imgout);
-
-    long nelement =
-        img1.md->nelement;
-
-#ifdef _OPENMP
-    #pragma omp parallel if (nelement > OMP_NELEMENT_LIMIT)
-    {
-        #pragma omp for simd
-#endif
-        for(long ii = 0;
-            ii < nelement; ii++)
-        {
-            imgout.im->array.CD[ii] =
-                pt2function(
-                    img1.im->array.CD[ii],
-                    img2.im->array.CD[ii]);
-        }
-#ifdef _OPENMP
-    }
-#endif
-    RegisterIMGID(&imgout, dcimg, dcnimg);
-    imgid_free(&img1);
-    imgid_free(&img2);
-    imgid_free(&imgout);
-    return RETURN_SUCCESS;
-}
+// Migrated to im_complex.c
 
 int arith_image_function_1f_1_IMGID(IMGID *imgin, double f1, IMGID *imgout, double (*pt2function)(double, double))
 {
@@ -1209,475 +1085,8 @@ int arith_image_function_1ff_1_inplace(const char *ID_name, double f1, double f2
 /* ---------------------------------------------------------- */
 /* Unary optimized: calls float/double math directly          */
 /* ---------------------------------------------------------- */
-#define ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(         \
-    name, funcname, funcname_f)                       \
-errno_t arith_image_##name##_optimized_IMGID(         \
-    IMGID *imgin, IMGID *imgout)                      \
-{                                                     \
-    DEBUG_TRACE_FSTART();                             \
-    if (imgin->im == NULL)                            \
-    {                                                 \
-        return RETURN_FAILURE;                        \
-    }                                                 \
-    imgid_ensure_output(imgin, imgout);               \
-    uint64_t nelement = imgout->md->nelement;         \
-    if (imgin->md->datatype == _DATATYPE_FLOAT        \
-        && imgout->mdt->datatype == _DATATYPE_FLOAT)  \
-    {                                                 \
-        float * MILK_RESTRICT p1 =                    \
-            MILK_ASSUME_ALIGNED(imgin->im->array.F);  \
-        float * MILK_RESTRICT po =                    \
-            MILK_ASSUME_ALIGNED(imgout->im->array.F); \
-        _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")     \
-        for (uint64_t i = 0; i < nelement; i++)       \
-        {                                             \
-            po[i] = funcname_f(p1[i]);                \
-        }                                             \
-    }                                                 \
-    else if (imgin->md->datatype == _DATATYPE_DOUBLE  \
-        && imgout->mdt->datatype == _DATATYPE_DOUBLE) \
-    {                                                 \
-        double *p1 = imgin->im->array.D;              \
-        double *po = imgout->im->array.D;             \
-        _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")     \
-        for (uint64_t i = 0; i < nelement; i++)       \
-        {                                             \
-            po[i] = funcname(p1[i]);                  \
-        }                                             \
-    }                                                 \
-    else                                              \
-    {                                                 \
-        arith_image_function_1_1_IMGID(               \
-            imgin, imgout, &P##name);                 \
-    }                                                 \
-    DEBUG_TRACE_FEXIT();                              \
-    return RETURN_SUCCESS;                            \
-}
+// Migrated to im_math_unary.c
 
-ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(acos, acos, acosf)
-ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(asin, asin, asinf)
-ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(atan, atan, atanf)
-ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(ceil, ceil, ceilf)
-ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(cos, cos, cosf)
-ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(cosh, cosh, coshf)
-ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(exp, exp, expf)
-ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(fabs, fabs, fabsf)
-ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(floor, floor, floorf)
-ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(ln, log, logf)
-ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(log, log10, log10f)
-ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(sqrt, sqrt, sqrtf)
-ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(sin, sin, sinf)
-ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(sinh, sinh, sinhf)
-ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(tan, tan, tanf)
-ARITH_UNARY_OPTIMIZED_FUNCTION_CALL(tanh, tanh, tanhf)
+// Migrated to im_math_basic.c
 
-#define ARITH_OPTIMIZED_FUNCTION(name, op) \
-errno_t arith_image_##name##_optimized_IMGID(IMGID *imgin1, IMGID *imgin2, IMGID *imgout) \
-{ \
-    DEBUG_TRACE_FSTART(); \
-    if (imgin1->im == NULL || imgin2->im == NULL) { return RETURN_FAILURE; } \
-    imgid_ensure_output(imgin1, imgout); \
-    uint64_t nelement = imgout->md->nelement; \
-    if(imgin1->md->datatype == _DATATYPE_FLOAT && imgin2->md->datatype == _DATATYPE_FLOAT && imgout->mdt->datatype == _DATATYPE_FLOAT) \
-    { \
-        float * MILK_RESTRICT p1 = MILK_ASSUME_ALIGNED(imgin1->im->array.F); \
-        float * MILK_RESTRICT p2 = MILK_ASSUME_ALIGNED(imgin2->im->array.F); \
-        float * MILK_RESTRICT po = MILK_ASSUME_ALIGNED(imgout->im->array.F); \
-        _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)") \
-        for(uint64_t i=0; i<nelement; i++) po[i] = p1[i] op p2[i]; \
-    } \
-    else if(imgin1->md->datatype == _DATATYPE_DOUBLE && imgin2->md->datatype == _DATATYPE_DOUBLE && imgout->mdt->datatype == _DATATYPE_DOUBLE) \
-    { \
-        double * MILK_RESTRICT p1 = MILK_ASSUME_ALIGNED(imgin1->im->array.D); \
-        double * MILK_RESTRICT p2 = MILK_ASSUME_ALIGNED(imgin2->im->array.D); \
-        double * MILK_RESTRICT po = MILK_ASSUME_ALIGNED(imgout->im->array.D); \
-        _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)") \
-        for(uint64_t i=0; i<nelement; i++) po[i] = p1[i] op p2[i]; \
-    } \
-    else \
-    { \
-        arith_image_function_2_1_IMGID(imgin1, imgin2, imgout, &P##name); \
-    } \
-    DEBUG_TRACE_FEXIT(); \
-    return RETURN_SUCCESS; \
-}
-
-ARITH_OPTIMIZED_FUNCTION(add, +)
-ARITH_OPTIMIZED_FUNCTION(sub, -)
-ARITH_OPTIMIZED_FUNCTION(mult, *)
-ARITH_OPTIMIZED_FUNCTION(div, /)
-
-#define ARITH_CST_OPTIMIZED_FUNCTION(name, op) \
-errno_t arith_image_cst##name##_optimized_IMGID(IMGID *imgin, double f1, IMGID *imgout) \
-{ \
-    DEBUG_TRACE_FSTART(); \
-    if (imgin->im == NULL) { return RETURN_FAILURE; } \
-    imgid_ensure_output(imgin, imgout); \
-    uint64_t nelement = imgout->md->nelement; \
-    if(imgin->md->datatype == _DATATYPE_FLOAT && imgout->mdt->datatype == _DATATYPE_FLOAT) \
-    { \
-        float * MILK_RESTRICT p1 = MILK_ASSUME_ALIGNED(imgin->im->array.F); \
-        float * MILK_RESTRICT po = MILK_ASSUME_ALIGNED(imgout->im->array.F); \
-        float cf1 = (float)f1; \
-        _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)") \
-        for(uint64_t i=0; i<nelement; i++) po[i] = p1[i] op cf1; \
-    } \
-    else if(imgin->md->datatype == _DATATYPE_DOUBLE && imgout->mdt->datatype == _DATATYPE_DOUBLE) \
-    { \
-        double *p1 = imgin->im->array.D; \
-        double *po = imgout->im->array.D; \
-        _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)") \
-        for(uint64_t i=0; i<nelement; i++) po[i] = p1[i] op f1; \
-    } \
-    else \
-    { \
-        arith_image_function_1f_1_IMGID(imgin, f1, imgout, &P##name); \
-    } \
-    DEBUG_TRACE_FEXIT(); \
-    return RETURN_SUCCESS; \
-}
-
-ARITH_CST_OPTIMIZED_FUNCTION(add, +)
-ARITH_CST_OPTIMIZED_FUNCTION(sub, -)
-ARITH_CST_OPTIMIZED_FUNCTION(mult, *)
-ARITH_CST_OPTIMIZED_FUNCTION(div, /)
-
-errno_t arith_image_cstpow_optimized_IMGID(IMGID *imgin, double f1, IMGID *imgout)
-{
-    DEBUG_TRACE_FSTART();
-    if (imgin->im == NULL) { return RETURN_FAILURE; }
-    imgid_ensure_output(imgin, imgout);
-    uint64_t nelement = imgout->md->nelement;
-    if(imgin->md->datatype == _DATATYPE_FLOAT && imgout->mdt->datatype == _DATATYPE_FLOAT)
-    {
-        float * MILK_RESTRICT p1 = MILK_ASSUME_ALIGNED(imgin->im->array.F);
-        float * MILK_RESTRICT po = MILK_ASSUME_ALIGNED(imgout->im->array.F);
-        float cf1 = (float)f1;
-        if (f1 == 0.0) {
-            _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")
-            for(uint64_t i=0; i<nelement; i++) po[i] = 1.0f;
-        } else if (f1 == 1.0) {
-            _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")
-            for(uint64_t i=0; i<nelement; i++) po[i] = p1[i];
-        } else if (f1 == 0.5) {
-            _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")
-            for(uint64_t i=0; i<nelement; i++) po[i] = sqrtf(p1[i]);
-        } else if (f1 == 2.0) {
-            _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")
-            for(uint64_t i=0; i<nelement; i++) po[i] = p1[i] * p1[i];
-        } else {
-            _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")
-            for(uint64_t i=0; i<nelement; i++) po[i] = powf(p1[i], cf1);
-        }
-    }
-    else if(imgin->md->datatype == _DATATYPE_DOUBLE && imgout->mdt->datatype == _DATATYPE_DOUBLE)
-    {
-        double * MILK_RESTRICT p1 = MILK_ASSUME_ALIGNED(imgin->im->array.D);
-        double * MILK_RESTRICT po = MILK_ASSUME_ALIGNED(imgout->im->array.D);
-        if (f1 == 0.0) {
-            _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")
-            for(uint64_t i=0; i<nelement; i++) po[i] = 1.0;
-        } else if (f1 == 1.0) {
-            _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")
-            for(uint64_t i=0; i<nelement; i++) po[i] = p1[i];
-        } else if (f1 == 0.5) {
-            _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")
-            for(uint64_t i=0; i<nelement; i++) po[i] = sqrt(p1[i]);
-        } else if (f1 == 2.0) {
-            _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")
-            for(uint64_t i=0; i<nelement; i++) po[i] = p1[i] * p1[i];
-        } else {
-            _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")
-            for(uint64_t i=0; i<nelement; i++) po[i] = pow(p1[i], f1);
-        }
-    }
-    else
-    {
-        arith_image_function_1f_1_IMGID(imgin, f1, imgout, &Ppow);
-    }
-    DEBUG_TRACE_FEXIT();
-    return RETURN_SUCCESS;
-}
-
-#define ARITH_OPTIMIZED_FUNCTION_CALL(name, funcname, funcname_f) \
-errno_t arith_image_##name##_optimized_IMGID(IMGID *imgin1, IMGID *imgin2, IMGID *imgout) \
-{ \
-    DEBUG_TRACE_FSTART(); \
-    if (imgin1->im == NULL || imgin2->im == NULL) { return RETURN_FAILURE; } \
-    imgid_ensure_output(imgin1, imgout); \
-    uint64_t nelement = imgout->md->nelement; \
-    if(imgin1->md->datatype == _DATATYPE_FLOAT && imgin2->md->datatype == _DATATYPE_FLOAT && imgout->mdt->datatype == _DATATYPE_FLOAT) \
-    { \
-        float * MILK_RESTRICT p1 = MILK_ASSUME_ALIGNED(imgin1->im->array.F); \
-        float * MILK_RESTRICT p2 = MILK_ASSUME_ALIGNED(imgin2->im->array.F); \
-        float * MILK_RESTRICT po = MILK_ASSUME_ALIGNED(imgout->im->array.F); \
-        _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)") \
-        for(uint64_t i=0; i<nelement; i++) po[i] = funcname_f(p1[i], p2[i]); \
-    } \
-    else if(imgin1->md->datatype == _DATATYPE_DOUBLE && imgin2->md->datatype == _DATATYPE_DOUBLE && imgout->mdt->datatype == _DATATYPE_DOUBLE) \
-    { \
-        double * MILK_RESTRICT p1 = MILK_ASSUME_ALIGNED(imgin1->im->array.D); \
-        double * MILK_RESTRICT p2 = MILK_ASSUME_ALIGNED(imgin2->im->array.D); \
-        double * MILK_RESTRICT po = MILK_ASSUME_ALIGNED(imgout->im->array.D); \
-        _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)") \
-        for(uint64_t i=0; i<nelement; i++) po[i] = funcname(p1[i], p2[i]); \
-    } \
-    else \
-    { \
-        arith_image_function_2_1_IMGID(imgin1, imgin2, imgout, &P##name); \
-    } \
-    DEBUG_TRACE_FEXIT(); \
-    return RETURN_SUCCESS; \
-}
-
-ARITH_OPTIMIZED_FUNCTION_CALL(pow, pow, powf)
-ARITH_OPTIMIZED_FUNCTION_CALL(fmod, fmod, fmodf)
-ARITH_OPTIMIZED_FUNCTION_CALL(minv, fmin, fminf)
-ARITH_OPTIMIZED_FUNCTION_CALL(maxv, fmax, fmaxf)
-
-/* ---------------------------------------------------------- */
-/* Binary optimized: comparison/logic with expression          */
-/* ---------------------------------------------------------- */
-#define ARITH_OPTIMIZED_FUNCTION_EXPR(               \
-    name, expr_f, expr_d)                             \
-errno_t arith_image_##name##_optimized_IMGID(         \
-    IMGID *imgin1,                                    \
-    IMGID *imgin2,                                    \
-    IMGID *imgout)                                    \
-{                                                     \
-    DEBUG_TRACE_FSTART();                             \
-    if (imgin1->im == NULL                            \
-        || imgin2->im == NULL)                        \
-    {                                                 \
-        return RETURN_FAILURE;                        \
-    }                                                 \
-    imgid_ensure_output(imgin1, imgout);              \
-    uint64_t nelement = imgout->md->nelement;         \
-    if (imgin1->md->datatype == _DATATYPE_FLOAT       \
-        && imgin2->md->datatype == _DATATYPE_FLOAT    \
-        && imgout->mdt->datatype == _DATATYPE_FLOAT)  \
-    {                                                 \
-        const float * MILK_RESTRICT p1 =              \
-            MILK_ASSUME_ALIGNED(                      \
-                imgin1->im->array.F);                 \
-        const float * MILK_RESTRICT p2 =              \
-            MILK_ASSUME_ALIGNED(                      \
-                imgin2->im->array.F);                 \
-        float * MILK_RESTRICT po =                    \
-            MILK_ASSUME_ALIGNED(                      \
-                imgout->im->array.F);                 \
-        _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")     \
-        for (uint64_t i = 0; i < nelement; i++)       \
-        {                                             \
-            po[i] = (expr_f);                         \
-        }                                             \
-    }                                                 \
-    else if (imgin1->md->datatype == _DATATYPE_DOUBLE \
-        && imgin2->md->datatype == _DATATYPE_DOUBLE   \
-        && imgout->mdt->datatype == _DATATYPE_DOUBLE) \
-    {                                                 \
-        const double * MILK_RESTRICT p1 =             \
-            MILK_ASSUME_ALIGNED(                      \
-                imgin1->im->array.D);                 \
-        const double * MILK_RESTRICT p2 =             \
-            MILK_ASSUME_ALIGNED(                      \
-                imgin2->im->array.D);                 \
-        double * MILK_RESTRICT po =                   \
-            MILK_ASSUME_ALIGNED(                      \
-                imgout->im->array.D);                 \
-        _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")     \
-        for (uint64_t i = 0; i < nelement; i++)       \
-        {                                             \
-            po[i] = (expr_d);                         \
-        }                                             \
-    }                                                 \
-    else                                              \
-    {                                                 \
-        arith_image_function_2_1_IMGID(               \
-            imgin1, imgin2, imgout, &P##name);        \
-    }                                                 \
-    DEBUG_TRACE_FEXIT();                              \
-    return RETURN_SUCCESS;                            \
-}
-
-ARITH_OPTIMIZED_FUNCTION_EXPR(
-    testlt,
-    (p1[i] < p2[i]) ? 1.0f : 0.0f,
-    (p1[i] < p2[i]) ? 1.0 : 0.0)
-ARITH_OPTIMIZED_FUNCTION_EXPR(
-    testmt,
-    (p1[i] >= p2[i]) ? 1.0f : 0.0f,
-    (p1[i] >= p2[i]) ? 1.0 : 0.0)
-ARITH_OPTIMIZED_FUNCTION_EXPR(
-    teste,
-    (p1[i] == p2[i]) ? 1.0f : 0.0f,
-    (p1[i] == p2[i]) ? 1.0 : 0.0)
-ARITH_OPTIMIZED_FUNCTION_EXPR(
-    testne,
-    (p1[i] != p2[i]) ? 1.0f : 0.0f,
-    (p1[i] != p2[i]) ? 1.0 : 0.0)
-ARITH_OPTIMIZED_FUNCTION_EXPR(
-    testle,
-    (p1[i] <= p2[i]) ? 1.0f : 0.0f,
-    (p1[i] <= p2[i]) ? 1.0 : 0.0)
-ARITH_OPTIMIZED_FUNCTION_EXPR(
-    testge,
-    (p1[i] >= p2[i]) ? 1.0f : 0.0f,
-    (p1[i] >= p2[i]) ? 1.0 : 0.0)
-ARITH_OPTIMIZED_FUNCTION_EXPR(
-    and,
-    ((p1[i] != 0.0f) && (p2[i] != 0.0f))
-        ? 1.0f : 0.0f,
-    ((p1[i] != 0.0) && (p2[i] != 0.0))
-        ? 1.0 : 0.0)
-ARITH_OPTIMIZED_FUNCTION_EXPR(
-    or,
-    ((p1[i] != 0.0f) || (p2[i] != 0.0f))
-        ? 1.0f : 0.0f,
-    ((p1[i] != 0.0) || (p2[i] != 0.0))
-        ? 1.0 : 0.0)
-
-/* ---------------------------------------------------------- */
-/* Const-scalar optimized: comparison/logic with expression    */
-/* ---------------------------------------------------------- */
-#define ARITH_CST_OPTIMIZED_FUNCTION_EXPR(            \
-    name, expr_f, expr_d)                             \
-errno_t arith_image_cst##name##_optimized_IMGID(      \
-    IMGID *imgin,                                     \
-    double f1,                                        \
-    IMGID *imgout)                                    \
-{                                                     \
-    DEBUG_TRACE_FSTART();                             \
-    if (imgin->im == NULL)                            \
-    {                                                 \
-        return RETURN_FAILURE;                        \
-    }                                                 \
-    imgid_ensure_output(imgin, imgout);               \
-    uint64_t nelement = imgout->md->nelement;         \
-    if (imgin->md->datatype == _DATATYPE_FLOAT        \
-        && imgout->mdt->datatype == _DATATYPE_FLOAT)  \
-    {                                                 \
-        const float * MILK_RESTRICT p1 =              \
-            MILK_ASSUME_ALIGNED(imgin->im->array.F);  \
-        float * MILK_RESTRICT po =                    \
-            MILK_ASSUME_ALIGNED(imgout->im->array.F); \
-        float cf1 = (float) f1;                       \
-        _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")     \
-        for (uint64_t i = 0; i < nelement; i++)       \
-        {                                             \
-            po[i] = (expr_f);                         \
-        }                                             \
-    }                                                 \
-    else if (imgin->md->datatype == _DATATYPE_DOUBLE  \
-        && imgout->mdt->datatype == _DATATYPE_DOUBLE) \
-    {                                                 \
-        const double * MILK_RESTRICT p1 =             \
-            MILK_ASSUME_ALIGNED(imgin->im->array.D);  \
-        double * MILK_RESTRICT po =                   \
-            MILK_ASSUME_ALIGNED(imgout->im->array.D); \
-        _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")     \
-        for (uint64_t i = 0; i < nelement; i++)       \
-        {                                             \
-            po[i] = (expr_d);                         \
-        }                                             \
-    }                                                 \
-    else                                              \
-    {                                                 \
-        arith_image_function_1f_1_IMGID(              \
-            imgin, f1, imgout, &P##name);             \
-    }                                                 \
-    DEBUG_TRACE_FEXIT();                              \
-    return RETURN_SUCCESS;                            \
-}
-
-ARITH_CST_OPTIMIZED_FUNCTION_EXPR(
-    testlt,
-    (p1[i] < cf1) ? 1.0f : 0.0f,
-    (p1[i] < f1) ? 1.0 : 0.0)
-ARITH_CST_OPTIMIZED_FUNCTION_EXPR(
-    testmt,
-    (p1[i] >= cf1) ? 1.0f : 0.0f,
-    (p1[i] >= f1) ? 1.0 : 0.0)
-ARITH_CST_OPTIMIZED_FUNCTION_EXPR(
-    teste,
-    (p1[i] == cf1) ? 1.0f : 0.0f,
-    (p1[i] == f1) ? 1.0 : 0.0)
-ARITH_CST_OPTIMIZED_FUNCTION_EXPR(
-    testne,
-    (p1[i] != cf1) ? 1.0f : 0.0f,
-    (p1[i] != f1) ? 1.0 : 0.0)
-ARITH_CST_OPTIMIZED_FUNCTION_EXPR(
-    testle,
-    (p1[i] <= cf1) ? 1.0f : 0.0f,
-    (p1[i] <= f1) ? 1.0 : 0.0)
-ARITH_CST_OPTIMIZED_FUNCTION_EXPR(
-    testge,
-    (p1[i] >= cf1) ? 1.0f : 0.0f,
-    (p1[i] >= f1) ? 1.0 : 0.0)
-ARITH_CST_OPTIMIZED_FUNCTION_EXPR(
-    and,
-    ((p1[i] != 0.0f) && (cf1 != 0.0f))
-        ? 1.0f : 0.0f,
-    ((p1[i] != 0.0) && (f1 != 0.0))
-        ? 1.0 : 0.0)
-ARITH_CST_OPTIMIZED_FUNCTION_EXPR(
-    or,
-    ((p1[i] != 0.0f) || (cf1 != 0.0f))
-        ? 1.0f : 0.0f,
-    ((p1[i] != 0.0) || (f1 != 0.0))
-        ? 1.0 : 0.0)
-
-/* ---------------------------------------------------------- */
-/* Unary optimized: comparison with expression                 */
-/* ---------------------------------------------------------- */
-#define ARITH_UNARY_OPTIMIZED_FUNCTION_EXPR(          \
-    name, expr_f, expr_d)                             \
-errno_t arith_image_##name##_optimized_IMGID(         \
-    IMGID *imgin, IMGID *imgout)                      \
-{                                                     \
-    DEBUG_TRACE_FSTART();                             \
-    if (imgin->im == NULL)                            \
-    {                                                 \
-        return RETURN_FAILURE;                        \
-    }                                                 \
-    imgid_ensure_output(imgin, imgout);               \
-    uint64_t nelement = imgout->md->nelement;         \
-    if (imgin->md->datatype == _DATATYPE_FLOAT        \
-        && imgout->mdt->datatype == _DATATYPE_FLOAT)  \
-    {                                                 \
-        const float * MILK_RESTRICT p1 =              \
-            MILK_ASSUME_ALIGNED(imgin->im->array.F);  \
-        float * MILK_RESTRICT po =                    \
-            MILK_ASSUME_ALIGNED(imgout->im->array.F); \
-        _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")     \
-        for (uint64_t i = 0; i < nelement; i++)       \
-        {                                             \
-            po[i] = (expr_f);                         \
-        }                                             \
-    }                                                 \
-    else if (imgin->md->datatype == _DATATYPE_DOUBLE  \
-        && imgout->mdt->datatype == _DATATYPE_DOUBLE) \
-    {                                                 \
-        const double * MILK_RESTRICT p1 =             \
-            MILK_ASSUME_ALIGNED(imgin->im->array.D);  \
-        double * MILK_RESTRICT po =                   \
-            MILK_ASSUME_ALIGNED(imgout->im->array.D); \
-        _Pragma("omp parallel for simd if (nelement > OMP_NELEMENT_LIMIT)")     \
-        for (uint64_t i = 0; i < nelement; i++)       \
-        {                                             \
-            po[i] = (expr_d);                         \
-        }                                             \
-    }                                                 \
-    else                                              \
-    {                                                 \
-        arith_image_function_1_1_IMGID(               \
-            imgin, imgout, &P##name);                 \
-    }                                                 \
-    DEBUG_TRACE_FEXIT();                              \
-    return RETURN_SUCCESS;                            \
-}
-
-ARITH_UNARY_OPTIMIZED_FUNCTION_EXPR(
-    positive,
-    (p1[i] > 0.0f) ? 1.0f : 0.0f,
-    (p1[i] > 0.0) ? 1.0 : 0.0)
+// Migrated to im_math_logic.c
